pra config bobot
