+++ app/Controllers/Bobot.php
@@ -4,14 +4,17 @@ namespace App\Controllers;
 
 use App\Models\BobotModel;
 use App\Models\PenilaianModel;
+use App\Models\StaseModel;
 
 class Bobot extends BaseController
 {
     protected $bobotModel;
+    protected $staseModel;
     protected $penilaianModel;
     public function __construct()
     {
         $this->bobotModel = new BobotModel();
+        $this->staseModel = new StaseModel();
         $this->penilaianModel = new PenilaianModel();
     }
     public function index()
@@ -19,9 +22,9 @@ class Bobot extends BaseController
         $currentPage = $this->request->getVar('page_bobot') ? $this->request->getVar('page_bobot') : 1;
         $keyword = $this->request->getVar('keyword');
         if ($keyword) {
-            $bobot = $this->bobotModel->getBobotSearch($keyword);
+            $bobot = $this->staseModel->getStaseSearch($keyword, 'ASC');
         } else {
-            $bobot = $this->bobotModel->getBobot();
+            $bobot = $this->staseModel->getStase('ASC');
         }
 
         $data = [
@@ -30,18 +33,23 @@ class Bobot extends BaseController
             'breadcrumb' => ['Setting', 'Bobot Nilai'],
             'bobot' => $bobot->paginate($this->numberPage, 'bobot'),
             'penilaian' => $this->penilaianModel->getPenilaian()->get()->getResult(),
-            'pager' => $this->bobotModel->pager,
+            'pager' => $this->staseModel->pager,
             'currentPage' => $currentPage,
             'numberPage' => $this->numberPage,
             'validation' => \Config\Services::validation(),
             'menu' => $this->fetchMenu()
         ];
-        // dd($data['penilaian']);
         return view('pages/bobot', $data);
     }
 
     public function savePenilaian($id)
     {
-        dd($_POST);
+        //cek jika id kosong
+
+
+        //cek jika penilaian tidak ada dipilih
+
+
+        // $data = array('' => , );
     }
 }
