+++ app/Helpers/rumkit_helper.php
@@ -19,3 +19,10 @@ function getUser($id)
     $result = $model->getProfile(['users.id' => $id])->getResult()[0];
     return $result;
 }
+
+function getUserId($id)
+{
+    $model = new \App\Models\UsersModel;
+    $result = $model->getSpecificUser(['users.id' => $id])->getResult()[0];
+    return $result;
+}
