+++ app/Models/StaseModel.php
@@ -11,19 +11,23 @@ class StaseModel extends Model
     protected $allowedFields = ['staseNama', 'staseJumlahWeek', 'staseType'];
     protected $returnType = 'object';
 
-    public function getStase()
+
+    public function getStase($fil = null)
     {
+        ($fil != null) ? $fil = $fil : $fil = 'DESC';
+
         $builder = $this->table('stase');
         $builder->select('*');
-        $builder->orderBy('stase.staseId', 'DESC');
+        $builder->orderBy('stase.staseId', $fil);
         return $builder;
     }
 
-    public function getStaseSearch($keyword)
+    public function getStaseSearch($keyword, $fil = null)
     {
+        ($fil != null) ? $fil = $fil : $fil = 'DESC';
         $builder = $this->table('stase');
         $builder->select('*');
-        $builder->orderBy('stase.staseId', 'DESC');
+        $builder->orderBy('stase.staseId', $fil);
         $builder->like('stase.staseNama', $keyword);
         return $builder;
     }
